requirements.txt
# Pythonの依存パッケージリスト
requests==2.31.0
aiohttp==3.9.3
python-dotenv==1.0.1
mysql-connector-python==8.3.0
SQLAlchemy==2.0.27
cryptography==42.0.2
bcrypt==4.1.2
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pandas==2.2.0
pydantic==2.6.1
playwright==1.41.2
flake8==7.0.0
safety==2.3.5
numpy==1.26.3